Authorization, Policies, Gates

+++ app/Http/Controllers/PostsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 //use Illuminate\Support\Facades\DB;
 
 class PostsController extends Controller
@@ -47,6 +48,7 @@ class PostsController extends Controller
      */
     public function create()
     {
+        //$this->authorize('posts.create');
         return view('posts.create');
     }
 
@@ -89,6 +91,10 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
         return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
     }
 
@@ -101,7 +107,15 @@ class PostsController extends Controller
      */
     public function update(StorePost $request, $id)
     {
+
         $post = BlogPost::findOrFail($id);
+
+//        if (Gate::denies('update-post', $post)) {
+//            abort(403, "You can't edit this blog post!");
+//        };
+
+        $this->authorize($post);
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -120,6 +134,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+//        if (Gate::denies('delete-post', $post)) {
+//            abort(403, "You can't delete this blog post!");
+//        };
+
+        $this->authorize( $post);
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
